wasm::checked

Struct Store

pub struct Store<'b, T: Config> {
    inner: Store<'b, T>,
    id: StoreId,
}

Fields
inner: Store<'b, T>id: StoreId

A unique identifier for this store. This is used to verify that stored objects belong to the current Store.

Implementations

impl<'b, T: Config> Store<'b, T>

pub fn new(user_data: T) -> Self

pub fn module_instantiate( &mut self, validation_info: &ValidationInfo<'b>, extern_vals: Vec<StoredExternVal>, maybe_fuel: Option<u64>, ) -> Result<StoredInstantiationOutcome, RuntimeError>

This is a safe variant of Store::module_instantiate_unchecked.

pub fn instance_export( &self, module_addr: Stored<ModuleAddr>, name: &str, ) -> Result<StoredExternVal, RuntimeError>

This is a safe variant of Store::$1.

pub unsafe fn func_alloc( &mut self, func_type: FuncType, host_func: fn(&mut T, Vec<Value>) -> Result<Vec<Value>, HaltExecutionError>, ) -> Stored<FuncAddr>

This is a safer variant of Store::func_alloc_unchecked. It is functionally equal, with the only difference being that this function returns a Stored<FuncAddr>.

Safety

The caller has to guarantee that if the Values returned from the given host function are references, their addresses came either from the host function arguments or from the current Store object.

See Store::func_alloc for more information.

pub fn func_type( &self, func_addr: Stored<FuncAddr>, ) -> Result<FuncType, RuntimeError>

This is a safe variant of Store::func_type_unchecked.

pub fn invoke( &mut self, func_addr: Stored<FuncAddr>, params: Vec<StoredValue>, maybe_fuel: Option<u64>, ) -> Result<StoredRunState<T>, RuntimeError>

This is a safe variant of Store::invoke_unchecked.

pub fn table_alloc( &mut self, table_type: TableType, ref: StoredRef, ) -> Result<Stored<TableAddr>, RuntimeError>

This is a safe variant of Store::table_alloc_unchecked.

pub fn table_type( &self, table_addr: Stored<TableAddr>, ) -> Result<TableType, RuntimeError>

This is a safe variant of Store::table_type_unchecked.

pub fn table_read( &self, table_addr: Stored<TableAddr>, i: u32, ) -> Result<StoredRef, RuntimeError>

This is a safe variant of Store::table_read_unchecked.

pub fn table_write( &mut self, table_addr: Stored<TableAddr>, i: u32, ref: StoredRef, ) -> Result<(), RuntimeError>

This is a safe variant of Store::table_write_unchecked.

pub fn table_size( &self, table_addr: Stored<TableAddr>, ) -> Result<u32, RuntimeError>

This is a safe variant of Store::table_size_unchecked.

pub fn mem_alloc(&mut self, mem_type: MemType) -> Stored<MemAddr>

This is a variant of Store::mem_alloc that returns a stored object.

pub fn mem_type( &self, mem_addr: Stored<MemAddr>, ) -> Result<MemType, RuntimeError>

This is a safe variant of Store::mem_type_unchecked.

pub fn mem_read( &self, mem_addr: Stored<MemAddr>, i: u32, ) -> Result<u8, RuntimeError>

This is a safe variant of Store::mem_read_unchecked.

pub fn mem_write( &mut self, mem_addr: Stored<MemAddr>, i: u32, byte: u8, ) -> Result<(), RuntimeError>

This is a safe variant of Store::mem_write_unchecked.

pub fn mem_size(&self, mem_addr: Stored<MemAddr>) -> Result<u32, RuntimeError>

This is a safe variant of Store::mem_size_unchecked.

pub fn mem_grow( &mut self, mem_addr: Stored<MemAddr>, n: u32, ) -> Result<(), RuntimeError>

This is a safe variant of Store::mem_grow_unchecked.

pub fn global_alloc( &mut self, global_type: GlobalType, val: StoredValue, ) -> Result<Stored<GlobalAddr>, RuntimeError>

This is a safe variant of Store::global_alloc_unchecked.

pub fn global_type( &self, global_addr: Stored<GlobalAddr>, ) -> Result<GlobalType, RuntimeError>

This is a safe variant of Store::global_type_unchecked.

pub fn global_read( &self, global_addr: Stored<GlobalAddr>, ) -> Result<StoredValue, RuntimeError>

This is a safe variant of Store::global_read_unchecked.

pub fn global_write( &mut self, global_addr: Stored<GlobalAddr>, val: StoredValue, ) -> Result<(), RuntimeError>

This is a safe variant of Store::global_write_unchecked.

pub fn create_resumable( &self, func_addr: Stored<FuncAddr>, params: Vec<StoredValue>, maybe_fuel: Option<u64>, ) -> Result<Stored<Resumable<T>>, RuntimeError>

This is a safe variant of Store::create_resumable_unchecked.

pub fn resume( &mut self, resumable: Stored<Resumable<T>>, ) -> Result<StoredRunState<T>, RuntimeError>

This is a safe variant of Store::resume_unchecked.

pub fn func_alloc_typed<Params: InteropValueList, Returns: InteropValueList>( &mut self, host_func: fn(&mut T, Vec<Value>) -> Result<Vec<Value>, HaltExecutionError>, ) -> Stored<FuncAddr>

This is a safer variant of Store::func_alloc_typed_unchecked. It is functionally equal, with the only difference being that this function returns a Stored<FuncAddr>.

Safety

The caller has to guarantee that if the Values returned from the given host function are references, their addresses came either from the host function arguments or from the current Store object.

See: Store::func_alloc_typed_unchecked for more information.

pub fn invoke_without_fuel( &mut self, func_addr: Stored<FuncAddr>, params: Vec<StoredValue>, ) -> Result<Vec<StoredValue>, RuntimeError>

This is a safe variant of Store::invoke_without_fuel_unchecked.

pub fn invoke_typed_without_fuel<Params: StoredInteropValueList, Returns: StoredInteropValueList>( &mut self, function: Stored<FuncAddr>, params: Params, ) -> Result<Returns, RuntimeError>

This is a safe variant of Store::invoke_typed_without_fuel_unchecked.

pub fn mem_access_mut_slice<R>( &self, memory: Stored<MemAddr>, accessor: impl FnOnce(&mut [u8]) -> R, ) -> Result<R, RuntimeError>

This is a safe variant of Store::mem_access_mut_slice_unchecked.