wasm::core::reader

Module types

Methods to read WASM Types from a WasmReader object.

See: https://webassembly.github.io/spec/core/binary/types.html

Modules

data

element

export

global

import

memarg

opcode

All opcodes, in alphanumerical order by their numeric (hex-)value

values

Methods to read basic WASM Values from a WasmReader object.

Structs

FuncType

https://webassembly.github.io/spec/core/binary/types.html#function-types

Limits

MemType

ResultType

https://webassembly.github.io/spec/core/binary/types.html#value-types

TableType

VecType ๐

https://webassembly.github.io/spec/core/binary/types.html#vector-types

Enums

BlockType

https://webassembly.github.io/spec/core/binary/instructions.html#binary-blocktype

ExternType

https://webassembly.github.io/spec/core/valid/types.html#external-types

NumType

https://webassembly.github.io/spec/core/binary/types.html#number-types

RefType

https://webassembly.github.io/spec/core/binary/types.html#reference-types

ValType

https://webassembly.github.io/spec/core/binary/types.html#reference-types TODO flatten NumType and RefType enums, as they are not used individually and wasmparser also does it.

Traits

ImportSubTypeRelation