wasm::core::reader::types::element

Function parse_validate_shortened_initializer_list

fn parse_validate_shortened_initializer_list(
    wasm: &mut WasmReader<'_>,
    num_funcs: usize,
    validation_context_refs: &mut BTreeSet<usize>,
) -> Result<ElemItems, ValidationError>

Parse and validate a vector of func_idx’s that reflect as the initializer list of an element segment in the form of ((ref.func func_idx) end) of in the abstract syntax. An expression of such form is valid and const with type funcref if C.funcs[func_idx] exists (see link for the definition of validation context) This codebase holds these as ElemItems::RefFuncs(Vec<u32>) to sidestep the need to show Span for each expression. Additionally inserts new items to the set validation_context_refs. validation_context_refs corresponds to C.refs in https://webassembly.github.io/spec/core/valid/conventions.html#context

Returns

Ok(ElemItems::RefFuncs(_)) corresponding to the parsed list if parsing & validating succeeds, Err(_) otherwise.