wasm

Module execution

Modules

assert_validated 🔒

Helpers for assertions due to prior validation of a WASM program.

config

const_interpreter_loop

error

interop

This module provides types, traits and impls to convert between Rust types and the Wasm Value type. Internally this module is not used, except for the top-level entry points for invocation.

interpreter_loop 🔒

This module solely contains the actual interpretation loop that matches instructions, interpreting the WASM bytecode

linker

little_endian 🔒

This module contains the definition and implementation of LittleEndianBytes, a trait to convert values (such as integers or floats) to bytes in little endian byter order

registry

resumable

store

value

value_stack

Structs

RuntimeInstance

Constants

DEFAULT_MODULE

The default module name if a RuntimeInstance was created using RuntimeInstance::new.

Functions

host_function_wrapper

Helper function to quickly construct host functions without worrying about wasm to Rust type conversion. For reading/writing user data into the current configuration, simply move user_data into the passed closure.