wasm::execution::checked

Function wrap_vec_elements

unsafe fn wrap_vec_elements<S: AbstractStored>(
    values: Vec<S::BareTy>,
    id: StoreId,
) -> Vec<S>

Helper method for associating every element in a Vec with a StoreId.

Safety

It must be guaranteed that all given elements come from the Store with the given StoreId.