wasm::execution::checked

Struct Linker

pub struct Linker {
    inner: Linker,
    store_id: Option<StoreId>,
}

Fields
inner: Linkerstore_id: Option<StoreId>

This is for the checked API which makes sure that all objects used originate from the same Store.

Initially the store id is None. Only when a store-specific object or a Store itself is used with a checked method, is this field set. Once initialized it is never written to again.

Implementations

impl Linker

pub fn new() -> Self

Creates a new Linker that is not yet associated to any specific Store.

pub fn define( &mut self, module_name: String, name: String, extern_val: StoredExternVal, ) -> Result<(), RuntimeError>

This is a safe variant of Linker::define_unchecked.

pub fn define_module_instance<T: Config>( &mut self, store: &Store<'_, T>, module_name: String, module: Stored<ModuleAddr>, ) -> Result<(), RuntimeError>

This is a safe variant of Linker::define_module_instance_unchecked.

pub fn get( &self, module_name: String, name: String, ) -> Result<StoredExternVal, RuntimeError>

This is a variant of Linker::get.

Interaction with unchecked API

This method is able to find externs defined through the unchecked define methods. However, for this to work, at least one of the following methods must have been called successfully: Linker::define, Linker::define_module_instance, Linker::module_instantiate. Otherwise, this method may spuriously return an error.

Therefore, it is advised against to mix the unchecked and checked API for a single Linker instance.

Errors

RuntimeError::LinkerNotYetAssociatedWithStoreId

RuntimeError::UnableToResolveExternLookup

pub fn instantiate_pre( &self, validation_info: &ValidationInfo<'_>, ) -> Result<Vec<StoredExternVal>, RuntimeError>

This is a variant of Linker::instantiate_pre.

Interaction with unchecked API

See Linker::get

Errors

RuntimeError::LinkerNotYetAssociatedWithStoreId

RuntimeError::UnableToResolveExternLookup

pub fn module_instantiate<'b, T: Config>( &mut self, store: &mut Store<'b, T>, validation_info: &ValidationInfo<'b>, maybe_fuel: Option<u64>, ) -> Result<StoredInstantiationOutcome, RuntimeError>

This is a safe variant of Linker::module_instantiate_unchecked.