wasm::execution::config

Trait Config

pub trait Config {
    // Provided method
    fn instruction_hook(&mut self, bytecode: &[u8], pc: usize) { ... }
}

Trait that allows user specified configuration for various items during interpretation. Additionally, the types implementing this trait can act as custom user data within an interpreter instance, passed along to each method of this trait and host functions whenever they are invoked.

The default implementation of all trait methods have the least overhead, i. e. most can be optimized out fully.

Provided Methods

fn instruction_hook(&mut self, bytecode: &[u8], pc: usize)

A hook which is called before every wasm instruction

This allows the most intricate insight into the interpreters behavior, at the cost of a hefty performance penalty