wasm::execution

Module execution_info

Structs

ExecutionInfo

ExecutionInfo is a compilation of relevant information needed by the interpreter loop. The lifetime annotation 'r represents that this structure needs to be valid at least as long as the RuntimeInstance that creates it.