wasm::execution::interop

Trait InteropValue

pub trait InteropValue
where
    Self: Copy + Debug + PartialEq + TryFrom<Value, Error = ValueTypeMismatchError>,
    Value: From<Self>,
{
    const TY: ValType;
}

An InteropValue is a Rust types that can be converted into a WASM Value. This trait is intended to simplify translation between Rust values and WASM values and thus is not used internally.

Required Associated Constants

const TY: ValType

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.