wasm::execution

Module interpreter_loop

This module solely contains the actual interpretation loop that matches instructions, interpreting the WASM bytecode

Note to Developer:

There must be only imports and one impl with one function (run) in it.

This module must only use RuntimeError and never Error.

Modules

control 🔒

dispatch_tables 🔒

memory 🔒

numeric 🔒

parametric 🔒

reference 🔒

table 🔒

variable 🔒

vector 🔒

Macros

define_instruction_fn 🔒

Structs

Args 🔒

Enums

InterpreterLoopOutcome

A non-error outcome of execution of the interpreter loop

Functions

calculate_mem_address 🔒

data_drop 🔒 ⚠

Safety

decrement_fuel 🔒

do_sidetable_control_transfer 🔒

elem_drop 🔒 ⚠

Safety

fc_extensions 🔒 ⚠

Safety

fd_extensions 🔒 ⚠

Safety

from_lanes 🔒

memory_init 🔒 ⚠

Safety

run 🔒 ⚠

Interprets wasm native functions. Wasm parameters and Wasm return values are passed on the stack. Returns Ok(ControlFlow::Continue(())) in case execution successfully terminates, Ok(Some(required_fuel)) if execution terminates due to insufficient fuel, indicating how much fuel is required to resume with required_fuel, and [Error::RuntimeError] otherwise.

table_init 🔒 ⚠

Safety

to_lanes 🔒

unset 🔒 ⚠

Safety

Type Aliases

InstructionHandlerFn 🔒