wasm::execution::interpreter_loop

Function run

pub(super) fn run<T: Config>(
    resumable: &mut Resumable,
    store: &mut Store<'_, T>,
) -> Result<Option<NonZeroU32>, RuntimeError>

Interprets wasm native functions. Wasm parameters and Wasm return values are passed on the stack. Returns Ok(None) in case execution successfully terminates, Ok(Some(required_fuel)) if execution terminates due to insufficient fuel, indicating how much fuel is required to resume with required_fuel, and [Error::RuntimeError] otherwise.