wasm::execution

Module store

Modules

addrs

Type definitions for addr types

instances 🔒

linear_memory 🔒

Structs

HaltExecutionError

A marker error for host functions to return, in case they want execution to be halted.

InstantiationOutcome

Represents a successful, possibly fueled instantiation of a module.

Store

The store represents all global state that can be manipulated by WebAssembly programs. It consists of the runtime representation of all instances of functions, tables, memories, and globals, element segments, and data segments that have been allocated during the life time of the abstract machine. https://webassembly.github.io/spec/core/exec/runtime.html#store

StoreId

A unique identifier for a specific Store

Enums

ExternVal

https://webassembly.github.io/spec/core/exec/runtime.html#external-values

Traits

ExternFilterable

common convention functions defined for lists of ExternVals, ExternTypes, Exports https://webassembly.github.io/spec/core/exec/runtime.html#conventions https://webassembly.github.io/spec/core/syntax/types.html#id3 https://webassembly.github.io/spec/core/syntax/modules.html?highlight=convention#id1