wasm::execution::store

Struct Store

pub struct Store<'b> {
    pub functions: Vec<FuncInst>,
    pub memories: Vec<MemInst>,
    pub globals: Vec<GlobalInst>,
    pub data: Vec<DataInst>,
    pub tables: Vec<TableInst>,
    pub elements: Vec<ElemInst>,
    pub modules: Vec<ModuleInst<'b>>,
    pub registry: Registry,
}

The store represents all global state that can be manipulated by WebAssembly programs. It consists of the runtime representation of all instances of functions, tables, memories, and globals, element segments, and data segments that have been allocated during the life time of the abstract machine. https://webassembly.github.io/spec/core/exec/runtime.html#store

Fields
functions: Vec<FuncInst>memories: Vec<MemInst>globals: Vec<GlobalInst>data: Vec<DataInst>tables: Vec<TableInst>elements: Vec<ElemInst>modules: Vec<ModuleInst<'b>>registry: Registry
Implementations

impl<'b> Store<'b>

pub fn add_module( &mut self, name: &str, validation_info: &ValidationInfo<'b>, ) -> CustomResult<()>

instantiates a validated module with validation_info as validation evidence with name name with the steps in https://webassembly.github.io/spec/core/exec/modules.html#instantiation this method roughly matches the suggested embedder functionmodule_instantiate https://webassembly.github.io/spec/core/appendix/embedding.html#modules except external values for module instantiation are retrieved from self.

fn alloc_func( &mut self, func: (usize, (Span, usize)), module_inst: &ModuleInst<'_>, module_addr: usize, ) -> usize

roughly matches https://webassembly.github.io/spec/core/exec/modules.html#functions with the addition of sidetable pointer to the input signature

fn alloc_table(&mut self, table_type: TableType, reff: Ref) -> usize

https://webassembly.github.io/spec/core/exec/modules.html#tables

fn alloc_mem(&mut self, mem_type: MemType) -> usize

https://webassembly.github.io/spec/core/exec/modules.html#memories

fn alloc_global(&mut self, global_type: GlobalType, val: Value) -> usize

https://webassembly.github.io/spec/core/exec/modules.html#globals

fn alloc_elem(&mut self, ref_type: RefType, refs: Vec<Ref>) -> usize

https://webassembly.github.io/spec/core/exec/modules.html#element-segments

fn alloc_data(&mut self, bytes: &[u8]) -> usize

https://webassembly.github.io/spec/core/exec/modules.html#data-segments

pub fn invoke<Param: InteropValueList, Returns: InteropValueList>( &mut self, func_addr: usize, params: Param, ) -> Result<Returns, RuntimeError>

pub fn invoke_dynamic( &mut self, func_addr: usize, params: Vec<Value>, ret_types: &[ValType], ) -> Result<Vec<Value>, RuntimeError>

pub fn invoke_dynamic_unchecked_return_ty( &mut self, func_addr: usize, params: Vec<Value>, ) -> Result<Vec<Value>, RuntimeError>