wasm::execution::value_stack

Struct CallFrame

pub(crate) struct CallFrame {
    pub return_func_addr: usize,
    pub return_addr: usize,
    pub value_stack_base_idx: usize,
    pub call_frame_base_idx: usize,
    pub return_value_count: usize,
    pub return_stp: usize,
}

The WASM spec calls this Activations, however it refers to the call frames of functions.

Fields
return_func_addr: usize

Store address of the function that called this CallFrame’s function

return_addr: usize

Value that the PC has to be set to when this function returns

value_stack_base_idx: usize

The index to the lowermost value in Stack::values belonging to this CallFrame’s stack

Values below this may still belong to this CallFrame, but they are locals. Consequently, this is the lowest index down to which the stack may be popped in this CallFrame. However, clearing up this CallFrame may require further popping, down to (and including!) the index stored in Self::call_frame_base_idx.

call_frame_base_idx: usize

The index to the lowermost value on Stack::values that belongs to this CallFrame

Clearing this CallFrame requires popping all elements on Stack::values down to (and including!) this index.

return_value_count: usize

Number of return values to retain on Stack::values when unwinding/popping a CallFrame

return_stp: usize