wasm

Module interop

This module provides types, traits and impls to convert between Rust types and the Wasm Value type. Internally this module is not used, except for the top-level entry points for invocation.

The main trait is InteropValue. It is implemented for all Rust types which can be converted into and from a Value through the From and TryFrom traits, respectively.

Then, the InteropValueList trait is a layer on top, allowing the same conversions but instead for tuples/lists containing multiple values.

Structs

RefExtern

RefFunc

Traits

InteropValue

An InteropValue is a Rust types that can be converted into a WASM Value. This trait is intended to simplify translation between Rust values and WASM values and thus is not used internally.

InteropValueList

An InteropValueList is an iterable list of InteropValues (i.e. Rust types that can be converted into WASM Values).