wasm::resumable

Struct ResumableRef

pub struct ResumableRef {
    dormitory: Weak<RwSpinLock<SlotMap<Resumable>>>,
    key: SlotMapKey<Resumable>,
}

Fields
dormitory: Weak<RwSpinLock<SlotMap<Resumable>>>key: SlotMapKey<Resumable>
Implementations

impl ResumableRef

pub fn access_fuel_mut<T, R>( &self, runtime_instance: &mut RuntimeInstance<'_, T>, f: impl FnOnce(&mut Option<u32>) -> R, ) -> Result<R, RuntimeError>

calls its argument f with a mutable reference of the fuel of the respective ResumableRef.

Fuel is stored as an Option<u32>, where None means that fuel is disabled and Some(x) means that x units of fuel is left. A ubiquitious use of this method would be using f to read or mutate the current fuel amount of the respective ResumableRef.

Example

use wasm::{resumable::RunState, validate, RuntimeInstance};
let wasm = [ 0x00, 0x61, 0x73, 0x6d, 0x01, 0x00, 0x00, 0x00,
            0x01, 0x04, 0x01, 0x60, 0x00, 0x00, 0x03, 0x02,
            0x01, 0x00, 0x07, 0x09, 0x01, 0x05, 0x6c, 0x6f,
            0x6f, 0x70, 0x73, 0x00, 0x00, 0x0a, 0x09, 0x01,
            0x07, 0x00, 0x03, 0x40, 0x0c, 0x00, 0x0b, 0x0b ];
// a simple module with a single function looping forever
let mut instance = RuntimeInstance::new_named((), "module", &validate(&wasm).unwrap()).unwrap();
let func_ref = instance.get_function_by_name("module", "loops").unwrap();
let resumable = instance.invoke_resumable(&func_ref, vec![], 0).unwrap();
match resumable {
    RunState::Resumable { resumable_ref, .. } => {
        // inspect and modify fuel content
        resumable_ref.access_fuel_mut(&mut instance, |x| { assert_eq!(*x, Some(0)); *x = None; }).unwrap();
    }
    _ => unreachable!("this function loops forever")
}

pub fn resume<T>( self, runtime_instance: &mut RuntimeInstance<'_, T>, ) -> Result<RunState, RuntimeError>

resumes execution of the resumable.