wasm::validation

Module code

Functions

read_declared_locals

read_instructions 🔒

validate_branch_and_generate_sidetable_entry 🔒

Validates a specific branch to some label by its index label_idx. Branches are generated by branching instructions and some can even generate multiple branches.

validate_code_section