wasm::validation::code

Function validate_branch_and_generate_sidetable_entry

fn validate_branch_and_generate_sidetable_entry(
    wasm: &WasmReader<'_>,
    label_idx: usize,
    stack: &mut ValidationStack,
    sidetable: &mut Vec<SidetableEntry>,
    unify_to_expected_types: bool,
) -> Result<(), ValidationError>

Validates a specific branch to some label by its index label_idx. Branches are generated by branching instructions and some can even generate multiple branches.