wasm::validation::globals

Function validate_global_section

pub(super) fn validate_global_section(
    wasm: &mut WasmReader<'_>,
    section_header: SectionHeader,
    imported_global_types: &[GlobalType],
    validation_context_refs: &mut BTreeSet<usize>,
    num_funcs: usize,
) -> Result<Vec<Global>, ValidationError>

Validate the global section.

The global section is a vector of global variables. Each Global variable is composed of a GlobalType and an initialization expression represented by a constant expression.

See read_constant_expression for more information.