wasm::validation::instructions::expressions

Function decode_and_validate_expr

pub unsafe fn decode_and_validate_expr<T: ValidationConfig>(
    wasm: &mut WasmDecoder<'_>,
    stack: &mut ValidationStack,
    sidetable: &mut Vec<SidetableEntry>,
    locals: &[ValType],
    c_globals: &IdxVec<GlobalIdx, Global>,
    fn_types: &IdxVec<TypeIdx, FuncType>,
    c_funcs: &IdxVec<FuncIdx, TypeIdx>,
    c_mems: &IdxVec<MemIdx, MemType>,
    data_count: Option<u32>,
    c_tables: &IdxVec<TableIdx, TableType>,
    c_elems: &IdxVec<ElemIdx, ElemType>,
    validation_context_refs: &BTreeSet<FuncIdx>,
    user_data: &mut T,
) -> Result<(), ValidationError>

Decodes an expression1 and validates it2.

Safety

The caller must ensure that all index values passed into this function are valid in the relevant IdxVec. The following table lists all index types and their respective IdxVec types:

Index	IdxVec

TypeIdx	IdxVec<TypeIdx, FuncType>

FuncIdx	IdxVec<FuncIdx, TypeIdx>

TableIdx	IdxVec<TableIdx, TableType>

WebAssembly Specification 2.0 - 5.4.9. Expressions. ↩

WebAssembly Specification 2.0 - 3.3.10. Expressions. ↩