wasm

Module value

Structs

ExternAddr

Represents the address of an external reference in the interpreter.

F32

F64

FuncAddr

Represents the address of a function within a WebAssembly module.

Enums

Ref

RefValueTy

Value

A value at runtime. This is essentially a duplicate of ValType just with additional values.