Trait wasm::core::little_endian::LittleEndianBytes

pub trait LittleEndianBytes<const N: usize> {
    // Required methods
    fn from_le_bytes(bytes: [u8; N]) -> Self;
    fn to_le_bytes(self) -> [u8; N];
}

Convert from and to the little endian byte representation of a value

N denotes the number of bytes required for the little endian representation

Required Methods

fn from_le_bytes(bytes: [u8; N]) -> Self

Convert from a byte array to Self

fn to_le_bytes(self) -> [u8; N]

Convert from self to a byte array

Object Safety

This trait is not object safe.