Module wasm::execution

Modules

assert_validated 🔒

Helpers for assertions due to prior validation of a WASM program.

const_interpreter_loop

execution_info 🔒

function_ref

hooks

interpreter_loop 🔒

This module solely contains the actual interpretation loop that matches instructions, interpreting the WASM bytecode

linear_memory 🔒

locals 🔒

lut 🔒

store 🔒

value

value_stack

Structs

RuntimeInstance

Constants

DEFAULT_MODULE

The default module name if a RuntimeInstance was created using RuntimeInstance::new.

Functions

get_address_offset 🔒

Used for getting the offset of an address.