wasm::execution

Module interpreter_loop

This module solely contains the actual interpretation loop that matches instructions, interpreting the WASM bytecode

Note to Developer:

There must be only imports and one impl with one function (run) in it.

This module must not use the Error enum.

Instead, only the RuntimeError enum shall be used

not to be confused with the Error enum’s Error::RuntimeError variant, which as per 2., we don not want

Functions

data_drop 🔒

do_sidetable_control_transfer 🔒

elem_drop 🔒

get_store_index 🔒

memory_init 🔒

run 🔒

Interprets a functions. Parameters and return values are passed on the stack.

table_init 🔒