wasm::execution

Module store

Structs

DataInst

ElemInst

https://webassembly.github.io/spec/core/exec/runtime.html#element-instances

GlobalInst

HaltExecutionError

A marker error for host functions to return, in case they want execution to be halted.

HostFuncInst

MemInst

ModuleInst

https://webassembly.github.io/spec/core/exec/runtime.html#module-instances

Store

The store represents all global state that can be manipulated by WebAssembly programs. It consists of the runtime representation of all instances of functions, tables, memories, and globals, element segments, and data segments that have been allocated during the life time of the abstract machine. https://webassembly.github.io/spec/core/exec/runtime.html#store

TableInst

WasmFuncInst

Enums

ExternVal

https://webassembly.github.io/spec/core/exec/runtime.html#external-values

FuncInst

Traits

ExternFilterable

common convention functions defined for lists of ExternVals, ExternTypes, Exports https://webassembly.github.io/spec/core/exec/runtime.html#conventions https://webassembly.github.io/spec/core/syntax/types.html#id3 https://webassembly.github.io/spec/core/syntax/modules.html?highlight=convention#id1