Trait wasm::execution::value::InteropValue

pub trait InteropValue:
    Copy
    + Debug
    + PartialEq {
    const TY: ValType;

    // Required methods
    fn into_value(self) -> Value;
    fn from_value(value: Value) -> Self;
}

An InteropValue is a Rust types that can be converted into a WASM Value. This trait is intended to simplify translation between Rust values and WASM values and thus is not used internally.

Required Associated Constants

const TY: ValType

Required Methods

fn into_value(self) -> Value

fn from_value(value: Value) -> Self

Object Safety

This trait is not object safe.