Struct wasm::execution::value_stack::Stack

pub(crate) struct Stack {
    values: Vec<Value>,
    frames: Vec<CallFrame>,
}

The stack at runtime containing

Values

Labels

Activations

See https://webassembly.github.io/spec/core/exec/runtime.html#stack

Fields
values: Vec<Value>

WASM values on the stack, i.e. the actual data that instructions operate on

frames: Vec<CallFrame>

Stack frames

Each time a function is called, a new frame is pushed, whenever a function returns, a frame is popped

Implementations

impl Stack

pub fn new() -> Self

pub fn pop_value(&mut self, ty: ValType) -> Value

Pop a value of the given ValType from the value stack

pub fn peek_value(&self, ty: ValType) -> Value

Copy a value of the given ValType from the value stack without removing it

pub fn push_value(&mut self, value: Value)

Push a value to the value stack

pub fn get_local(&mut self, idx: usize)

Copy a local variable to the top of the value stack

pub fn set_local(&mut self, idx: usize)

Pop value from the top of the value stack, writing it to the given local

pub fn tee_local(&mut self, idx: usize)

Copy value from top of the value stack to the given local

pub fn current_stackframe(&self) -> &CallFrame

Get a shared reference to the current CallFrame

pub fn current_stackframe_mut(&mut self) -> &mut CallFrame

Get a mutable reference to the current CallFrame

pub fn pop_stackframe(&mut self) -> usize

Pop a CallFrame from the call stack, returning the return address

pub fn push_stackframe( &mut self, func_idx: usize, func_ty: &FuncType, locals: Locals, return_addr: usize )

Push a stackframe to the call stack

Takes the current Self::values’s length as CallFrame::value_stack_base_idx.

pub fn callframe_count(&self) -> usize

Returns how many stackframes are on the stack, in total.

pub fn pop_tail_iter(&mut self, n: usize) -> Drain<'_, Value>

Pop n elements from the value stack’s tail as an iterator, with the first element being closest to the bottom of the value stack

Note that this is providing the values in reverse order compared to popping n values (which would yield the element closest to the top of the value stack first).

pub fn clear_callframe_values(&mut self)

Clear all of the values pushed to the value stack by the current stack frame