wasm::resumable

Enum RunState

pub enum RunState {
    Finished {
        values: Vec<Value>,
        maybe_remaining_fuel: Option<u32>,
    },
    Resumable {
        resumable_ref: ResumableRef,
        required_fuel: NonZeroU32,
    },
}

Represents the state of a possibly interrupted resumable.

Variants

Finished

represents a resumable that has executed completely with return values values and possibly remaining fuel maybe_remaining_fuel (has Some(remaining_fuel) for fuel-metered operations and None otherwise)

Fields

values: Vec<Value>

maybe_remaining_fuel: Option<u32>

Resumable

represents a resumable that has ran out of fuel during execution, missing at least required_fuel units of fuel to continue further execution.

Fields

resumable_ref: ResumableRef

required_fuel: NonZeroU32