wasm

Struct RuntimeInstance

pub struct RuntimeInstance<'b, T = (), H = EmptyHookSet>
where
    H: HookSet + Debug,
{
    pub hook_set: H,
    pub store: Store<'b, T>,
}

Fields
hook_set: Hstore: Store<'b, T>
Implementations

impl<'b, T> RuntimeInstance<'b, T, EmptyHookSet>

pub fn new(user_data: T) -> Self

pub fn new_with_default_module( user_data: T, validation_info: &ValidationInfo<'b>, ) -> Result<Self, RuntimeError>

pub fn new_named( user_data: T, module_name: &str, validation_info: &ValidationInfo<'b>, ) -> Result<Self, RuntimeError>

impl<'b, T, H> RuntimeInstance<'b, T, H>
where H: HookSet + Debug,

pub fn add_module( &mut self, module_name: &str, validation_info: &ValidationInfo<'b>, ) -> Result<(), RuntimeError>

pub fn new_with_hooks(user_data: T, hook_set: H) -> Self

pub fn get_function_by_name( &self, module_name: &str, function_name: &str, ) -> Result<FunctionRef, RuntimeError>

pub fn get_function_by_index( &self, module_addr: usize, function_idx: usize, ) -> Result<FunctionRef, RuntimeError>

pub fn invoke_typed<Params: InteropValueList, Returns: InteropValueList>( &mut self, function_ref: &FunctionRef, params: Params, ) -> Result<Returns, RuntimeError>

Invokes a function with the given parameters of type Param, and return types of type Returns.

pub fn invoke( &mut self, function_ref: &FunctionRef, params: Vec<Value>, ) -> Result<Vec<Value>, RuntimeError>

Invokes a function with the given parameters. The return types depend on the function signature.

pub fn add_host_function_typed<Params: InteropValueList, Returns: InteropValueList>( &mut self, module_name: &str, name: &str, host_func: fn(&mut T, Vec<Value>) -> Vec<Value>, ) -> Result<FunctionRef, RuntimeError>

Adds a host function under module namespace module_name with name name. roughly similar to func_alloc in https://webassembly.github.io/spec/core/appendix/embedding.html#functions except the host function is made visible to other modules through these names.

pub fn add_host_function( &mut self, module_name: &str, name: &str, host_func_ty: FuncType, host_func: fn(&mut T, Vec<Value>) -> Vec<Value>, ) -> Result<FunctionRef, RuntimeError>

pub fn user_data(&self) -> &T

pub fn user_data_mut(&mut self) -> &mut T