Struct wasm::RuntimeInstance

pub struct RuntimeInstance<'b, H = EmptyHookSet>
where
    H: HookSet,
{
    pub wasm_bytecode: &'b [u8],
    types: Vec<FuncType>,
    exports: Vec<Export>,
    store: Store,
    pub hook_set: H,
}

Fields
wasm_bytecode: &'b [u8]types: Vec<FuncType>exports: Vec<Export>store: Storehook_set: H
Implementations

impl<'b> RuntimeInstance<'b, EmptyHookSet>

pub fn new(validation_info: &ValidationInfo<'b>) -> Result<Self, RuntimeError>

impl<'b, H> RuntimeInstance<'b, H>
where H: HookSet,

pub fn new_with_hooks( validation_info: &ValidationInfo<'b>, hook_set: H ) -> Result<Self, RuntimeError>

pub fn invoke_named<Param: InteropValueList, Returns: InteropValueList>( &mut self, func_name: &str, param: Param ) -> Result<Returns, RuntimeError>

pub fn invoke_func<Param: InteropValueList, Returns: InteropValueList>( &mut self, func_idx: usize, params: Param ) -> Result<Returns, RuntimeError>

Can only invoke functions with signature [t1] -> [t2] as of now.

pub fn invoke_dynamic( &mut self, func_idx: usize, params: Vec<Value>, ret_types: &[ValType] ) -> Result<Vec<Value>, RuntimeError>

Invokes a function with the given parameters, and return types which are not known at compile time.

fn init_store(validation_info: &ValidationInfo<'_>) -> Store