wasm

Module value

Structs

ExternAddr

The WebAssembly specification defines an externaddr as an address to an “external” type, i.e. is a type which is managed by the embedder. For this interpreter the task of managing external objects and relating them to addresses is handed off to the user, which means that an ExternAddr can simply be seen as an integer that is opaque to Wasm code without any meaning assigned to it.

F32

F64

ValueTypeMismatchError

An error used in all TryFrom<Value> implementations for Rust types (i32, F32, Ref, …)

Enums

Ref

Value

A value at runtime. This is essentially a duplicate of ValType just with additional values.